backend/django_app/pyproject.toml
[tool.poetry]
name = "django-app"
version = "0.1.0"
description = ""
authors = ["Bartosz Kozłowski <bartoszkozlowski@bluerider.software>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
django = "4.2.8"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
requests = "2.31.0"
beautifulsoup4 = "4.12.2"
celery = "5.3.4"
django-celery-beat = "2.5.0"
django-celery-results = "2.5.1"
redis = "4.5.5"
django-redis = "5.2.0"
django-solo = "2.0.0"
djangorestframework = "^3.14.0"
langchain = "^0.2.1"
qdrant-client = "^1.6.9"
openai = "^1.3.7"
tiktoken = "^0.5.2"
django-geojson = "4.0.0"
django-cors-headers = "^4.3.1"
pillow = "^10.1.0"
djangorestframework-simplejwt = "^5.3.1"
django-storages = "^1.14.3"
boto3 = "^1.34.111"
langchain-community = "^0.2.1"
pypdf = "^4.2.0"
dspy = "^0.1.5"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
django-stubs = "^4.2.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
line_length = 100
include_trailing_comma = true
skip = ["build", ".venv", ".serverless", "node_modules", ".vscode"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | \.vscode
    | .pyenv
    | .serverless
    | node_modules
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.mypy]
python_version = "3.11"
